requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.10
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
virtualenv==20.25.1
